feat: add dedicated day view
